# Reuse single row block writer