--- a/core/trino-spi/src/main/java/io/trino/spi/block/AbstractSingleRowBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/AbstractSingleRowBlock.java
@@ -21,13 +21,6 @@ import java.util.List;
 public abstract class AbstractSingleRowBlock
         implements Block
 {
-    protected final int rowIndex;
-
-    protected AbstractSingleRowBlock(int rowIndex)
-    {
-        this.rowIndex = rowIndex;
-    }
-
     @Override
     public final List<Block> getChildren()
     {
@@ -37,6 +30,8 @@ public abstract class AbstractSingleRowBlock
     abstract Block[] getRawFieldBlocks();
 
     protected abstract Block getRawFieldBlock(int fieldIndex);
+
+    protected abstract int getRowIndex();
 
     private void checkFieldIndex(int position)
     {
@@ -49,119 +44,119 @@ public abstract class AbstractSingleRowBlock
     public boolean isNull(int position)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).isNull(rowIndex);
+        return getRawFieldBlock(position).isNull(getRowIndex());
     }
 
     @Override
     public byte getByte(int position, int offset)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getByte(rowIndex, offset);
+        return getRawFieldBlock(position).getByte(getRowIndex(), offset);
     }
 
     @Override
     public short getShort(int position, int offset)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getShort(rowIndex, offset);
+        return getRawFieldBlock(position).getShort(getRowIndex(), offset);
     }
 
     @Override
     public int getInt(int position, int offset)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getInt(rowIndex, offset);
+        return getRawFieldBlock(position).getInt(getRowIndex(), offset);
     }
 
     @Override
     public long getLong(int position, int offset)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getLong(rowIndex, offset);
+        return getRawFieldBlock(position).getLong(getRowIndex(), offset);
     }
 
     @Override
     public Slice getSlice(int position, int offset, int length)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getSlice(rowIndex, offset, length);
+        return getRawFieldBlock(position).getSlice(getRowIndex(), offset, length);
     }
 
     @Override
     public int getSliceLength(int position)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getSliceLength(rowIndex);
+        return getRawFieldBlock(position).getSliceLength(getRowIndex());
     }
 
     @Override
     public int compareTo(int position, int offset, int length, Block otherBlock, int otherPosition, int otherOffset, int otherLength)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).compareTo(rowIndex, offset, length, otherBlock, otherPosition, otherOffset, otherLength);
+        return getRawFieldBlock(position).compareTo(getRowIndex(), offset, length, otherBlock, otherPosition, otherOffset, otherLength);
     }
 
     @Override
     public boolean bytesEqual(int position, int offset, Slice otherSlice, int otherOffset, int length)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).bytesEqual(rowIndex, offset, otherSlice, otherOffset, length);
+        return getRawFieldBlock(position).bytesEqual(getRowIndex(), offset, otherSlice, otherOffset, length);
     }
 
     @Override
     public int bytesCompare(int position, int offset, int length, Slice otherSlice, int otherOffset, int otherLength)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).bytesCompare(rowIndex, offset, length, otherSlice, otherOffset, otherLength);
+        return getRawFieldBlock(position).bytesCompare(getRowIndex(), offset, length, otherSlice, otherOffset, otherLength);
     }
 
     @Override
     public void writeBytesTo(int position, int offset, int length, BlockBuilder blockBuilder)
     {
         checkFieldIndex(position);
-        getRawFieldBlock(position).writeBytesTo(rowIndex, offset, length, blockBuilder);
+        getRawFieldBlock(position).writeBytesTo(getRowIndex(), offset, length, blockBuilder);
     }
 
     @Override
     public boolean equals(int position, int offset, Block otherBlock, int otherPosition, int otherOffset, int length)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).equals(rowIndex, offset, otherBlock, otherPosition, otherOffset, length);
+        return getRawFieldBlock(position).equals(getRowIndex(), offset, otherBlock, otherPosition, otherOffset, length);
     }
 
     @Override
     public long hash(int position, int offset, int length)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).hash(rowIndex, offset, length);
+        return getRawFieldBlock(position).hash(getRowIndex(), offset, length);
     }
 
     @Override
     public <T> T getObject(int position, Class<T> clazz)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getObject(rowIndex, clazz);
+        return getRawFieldBlock(position).getObject(getRowIndex(), clazz);
     }
 
     @Override
     public void writePositionTo(int position, BlockBuilder blockBuilder)
     {
         checkFieldIndex(position);
-        getRawFieldBlock(position).writePositionTo(rowIndex, blockBuilder);
+        getRawFieldBlock(position).writePositionTo(getRowIndex(), blockBuilder);
     }
 
     @Override
     public Block getSingleValueBlock(int position)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getSingleValueBlock(rowIndex);
+        return getRawFieldBlock(position).getSingleValueBlock(getRowIndex());
     }
 
     @Override
     public long getEstimatedDataSizeForStats(int position)
     {
         checkFieldIndex(position);
-        return getRawFieldBlock(position).getEstimatedDataSizeForStats(rowIndex);
+        return getRawFieldBlock(position).getEstimatedDataSizeForStats(getRowIndex());
     }
 
     @Override