--- a/core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlock.java
@@ -27,10 +27,11 @@ public class SingleRowBlock
     private static final int INSTANCE_SIZE = ClassLayout.parseClass(SingleRowBlock.class).instanceSize();
 
     private final Block[] fieldBlocks;
+    private final int rowIndex;
 
     SingleRowBlock(int rowIndex, Block[] fieldBlocks)
     {
-        super(rowIndex);
+        this.rowIndex = rowIndex;
         this.fieldBlocks = fieldBlocks;
     }
 
@@ -62,7 +63,7 @@ public class SingleRowBlock
     {
         long sizeInBytes = 0;
         for (int i = 0; i < fieldBlocks.length; i++) {
-            sizeInBytes += getRawFieldBlock(i).getRegionSizeInBytes(rowIndex, 1);
+            sizeInBytes += getRawFieldBlock(i).getRegionSizeInBytes(getRowIndex(), 1);
         }
         return sizeInBytes;
     }
@@ -92,6 +93,7 @@ public class SingleRowBlock
         return SingleRowBlockEncoding.NAME;
     }
 
+    @Override
     public int getRowIndex()
     {
         return rowIndex;
@@ -122,6 +124,6 @@ public class SingleRowBlock
             // All blocks are already loaded
             return this;
         }
-        return new SingleRowBlock(rowIndex, loadedFieldBlocks);
+        return new SingleRowBlock(getRowIndex(), loadedFieldBlocks);
     }
 }