--- a/core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlockWriter.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlockWriter.java
@@ -24,24 +24,17 @@ public class SingleRowBlockWriter
         extends AbstractSingleRowBlock
         implements BlockBuilder
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(SingleRowBlockWriter.class).instanceSize();
+    public static final int INSTANCE_SIZE = ClassLayout.parseClass(SingleRowBlockWriter.class).instanceSize();
 
     private final BlockBuilder[] fieldBlockBuilders;
-    private final long initialBlockBuilderSize;
-    private int positionsWritten;
 
     private int currentFieldIndexToWrite;
+    private int rowIndex = -1;
     private boolean fieldBlockBuilderReturned;
 
-    SingleRowBlockWriter(int rowIndex, BlockBuilder[] fieldBlockBuilders)
+    SingleRowBlockWriter(BlockBuilder[] fieldBlockBuilders)
     {
-        super(rowIndex);
         this.fieldBlockBuilders = fieldBlockBuilders;
-        long initialBlockBuilderSize = 0;
-        for (BlockBuilder fieldBlockBuilder : fieldBlockBuilders) {
-            initialBlockBuilderSize += fieldBlockBuilder.getSizeInBytes();
-        }
-        this.initialBlockBuilderSize = initialBlockBuilderSize;
     }
 
     /**
@@ -75,13 +68,23 @@ public class SingleRowBlockWriter
     }
 
     @Override
+    protected int getRowIndex()
+    {
+        return rowIndex;
+    }
+
+    @Override
     public long getSizeInBytes()
     {
         long currentBlockBuilderSize = 0;
+         /*
+          * We need to use subtraction in order to compute size because getRegionSizeInBytes(0, 1)
+          * returns non-zero result even if field block builder has no position appended
+          */
         for (BlockBuilder fieldBlockBuilder : fieldBlockBuilders) {
-            currentBlockBuilderSize += fieldBlockBuilder.getSizeInBytes();
+            currentBlockBuilderSize += fieldBlockBuilder.getSizeInBytes() - fieldBlockBuilder.getRegionSizeInBytes(0, rowIndex);
         }
-        return currentBlockBuilderSize - initialBlockBuilderSize;
+        return currentBlockBuilderSize;
     }
 
     @Override
@@ -189,7 +192,6 @@ public class SingleRowBlockWriter
     private void entryAdded()
     {
         currentFieldIndexToWrite++;
-        positionsWritten++;
     }
 
     @Override
@@ -198,7 +200,7 @@ public class SingleRowBlockWriter
         if (fieldBlockBuilderReturned) {
             throw new IllegalStateException("field block builder has been returned");
         }
-        return positionsWritten;
+        return currentFieldIndexToWrite;
     }
 
     @Override
@@ -228,6 +230,24 @@ public class SingleRowBlockWriter
         else {
             return format("SingleRowBlockWriter{numFields=%d, fieldBlockBuilderReturned=true}", fieldBlockBuilders.length);
         }
+    }
+
+    void setRowIndex(int rowIndex)
+    {
+        if (this.rowIndex != -1) {
+            throw new IllegalStateException("SingleRowBlockWriter should be reset before usage");
+        }
+        this.rowIndex = rowIndex;
+    }
+
+    void reset()
+    {
+        if (this.rowIndex == -1) {
+            throw new IllegalStateException("SingleRowBlockWriter is already reset");
+        }
+        this.rowIndex = -1;
+        this.currentFieldIndexToWrite = 0;
+        this.fieldBlockBuilderReturned = false;
     }
 
     private void checkFieldIndexToWrite()