* Add benchmark BenchmarkRowBlockBuilder to bench RowBlockBuilder.beginBlockEntry()
  * Reuse SingleRowBlockWriter instance instead of creating it every time

```
         Benchmark                                                                           (rows)  (typesLength)  Mode  Cnt       Score       Error   Units
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              2  avgt    5       0.605 ±     0.066   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              2  avgt    5       0.508 ±     0.067   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              2  avgt    5    1912.558 ±    65.766  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              2  avgt    5    1311.819 ±     0.979  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              2  avgt    5    1273.917 ±    95.104    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              2  avgt    5     734.502 ±    97.158    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              2  avgt    5     612.496 ±    61.736  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              2  avgt    5       5.711 ±     0.004  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              2  avgt    5     408.120 ±    67.460    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              2  avgt    5       3.198 ±     0.423    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              2  avgt    5    1320.322 ±     0.742  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              2  avgt    5    1320.375 ±     0.986  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              2  avgt    5     879.617 ±    95.724    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              2  avgt    5     739.293 ±    97.792    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              2  avgt    5       0.381 ±     0.001  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              2  avgt    5       0.381 ±     0.001  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              2  avgt    5       0.254 ±     0.028    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              2  avgt    5       0.213 ±     0.028    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              2  avgt    5      40.000              counts
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              2  avgt    5      40.000              counts
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              2  avgt    5      67.000                  ms
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              2  avgt    5      53.000                  ms
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              4  avgt    5       1.137 ±     0.194   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              4  avgt    5       0.965 ±     0.247   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              4  avgt    5    1446.003 ±   745.039  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              4  avgt    5    1557.988 ±     9.044  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              4  avgt    5    1809.210 ±   963.746    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              4  avgt    5    1655.344 ±   412.794    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              4  avgt    5     327.684 ±   274.993  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              4  avgt    5       9.588 ±     0.673  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              4  avgt    5     410.905 ±   365.540    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              4  avgt    5      10.185 ±     2.395    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              4  avgt    5    1141.099 ±   474.297  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              4  avgt    5    1571.807 ±     9.125  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              4  avgt    5    1427.140 ±   594.993    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              4  avgt    5    1670.026 ±   416.455    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              4  avgt    5       0.753 ±     0.073  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              4  avgt    5       0.761 ±     0.004  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              4  avgt    5       0.942 ±     0.185    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              4  avgt    5       0.809 ±     0.202    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              4  avgt    5      45.000              counts
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              4  avgt    5      65.000              counts
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              4  avgt    5      63.000                  ms
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              4  avgt    5      80.000                  ms
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              8  avgt    5       2.262 ±     0.389   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                       10              8  avgt    5       1.747 ±     0.037   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              8  avgt    5    1462.316 ±    28.010  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                        10              8  avgt    5    1301.432 ±     1.075  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              8  avgt    5    3641.248 ±   555.241    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                   10              8  avgt    5    2503.436 ±    52.911    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              8  avgt    5     172.422 ±    24.232  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space               10              8  avgt    5      17.592 ±     0.655  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              8  avgt    5     429.066 ±    58.763    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm          10              8  avgt    5      33.841 ±     1.794    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              8  avgt    5     881.525 ±     1.012  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                  10              8  avgt    5     881.129 ±     0.729  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              8  avgt    5    2195.422 ±   376.088    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm             10              8  avgt    5    1694.941 ±    35.824    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              8  avgt    5       1.523 ±     0.002  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space           10              8  avgt    5       1.523 ±     0.001  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              8  avgt    5       3.793 ±     0.650    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm      10              8  avgt    5       2.930 ±     0.062    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              8  avgt    5     110.000              counts
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                             10              8  avgt    5     115.000              counts
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              8  avgt    5     114.000                  ms
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                              10              8  avgt    5     114.000                  ms
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              2  avgt    5       5.616 ±     0.270   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              2  avgt    5       5.019 ±     0.053   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              2  avgt    5    2009.009 ±    29.691  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              2  avgt    5    1361.696 ±     1.348  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              2  avgt    5   12424.520 ±   406.403    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              2  avgt    5    7530.159 ±    87.250    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              2  avgt    5     656.633 ±    46.408  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              2  avgt    5       6.165 ±     0.656  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              2  avgt    5    4060.902 ±   315.492    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              2  avgt    5      34.095 ±     3.841    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              2  avgt    5    1372.809 ±     1.067  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              2  avgt    5    1371.954 ±     1.360  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm            100              2  avgt    5    8490.338 ±   409.582    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm            100              2  avgt    5    7586.884 ±    87.913    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space          100              2  avgt    5       0.381 ±     0.001  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space          100              2  avgt    5       0.381 ±     0.001  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm     100              2  avgt    5       2.355 ±     0.114    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm     100              2  avgt    5       2.105 ±     0.024    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                            100              2  avgt    5      40.000              counts
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                            100              2  avgt    5      45.000              counts
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                             100              2  avgt    5      65.000                  ms
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                             100              2  avgt    5      59.000                  ms
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              4  avgt    5      11.773 ±     1.474   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              4  avgt    5       9.492 ±     0.442   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              4  avgt    5    1388.038 ±    38.807  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              4  avgt    5    1617.491 ±     1.035  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              4  avgt    5   17995.007 ±  1749.722    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              4  avgt    5   16916.095 ±   781.269    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              4  avgt    5     320.579 ±    56.247  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              4  avgt    5       9.966 ±     0.653  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              4  avgt    5    4158.110 ±  1011.443    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              4  avgt    5     104.242 ±    10.990    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              4  avgt    5    1094.612 ±     1.369  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              4  avgt    5    1634.544 ±     1.046  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm            100              4  avgt    5   14193.565 ±  1774.536    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm            100              4  avgt    5   17094.442 ±   789.529    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space          100              4  avgt    5       0.761 ±     0.001  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space          100              4  avgt    5       0.761 ±     0.001  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm     100              4  avgt    5       9.874 ±     1.234    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Survivor_Space.norm     100              4  avgt    5       7.956 ±     0.367    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                            100              4  avgt    5      65.000              counts
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.count                            100              4  avgt    5      65.000              counts
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                             100              4  avgt    5      88.000                  ms
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.time                             100              4  avgt    5      83.000                  ms
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              8  avgt    5      24.495 ±     1.330   us/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry                                      100              8  avgt    5      17.741 ±     0.576   us/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              8  avgt    5    1436.307 ±    74.508  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate                       100              8  avgt    5    1350.509 ±     3.310  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              8  avgt    5   38649.974 ±     0.001    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.alloc.rate.norm                  100              8  avgt    5   26394.021 ±   833.587    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              8  avgt    5     163.985 ±     8.507  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space              100              8  avgt    5      17.579 ±     0.646  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              8  avgt    5    4412.724 ±     0.001    B/op
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Eden_Space.norm         100              8  avgt    5     343.561 ±    13.530    B/op
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              8  avgt    5     877.247 ±    45.507  MB/sec
(after)  BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen                 100              8  avgt    5     920.445 ±     2.256  MB/sec
(before) BenchmarkRowBlockBuilder.benchmarkBeginBlockEntry:·gc.churn.G1_Old_Gen.norm            100              8  avgt    5   23606.071 ±     0.027    B/op
…